Spraycan is fully functional. YEEEEEAAAAH HAHAHAHAHA IT WORKSSSSSS
It's so satisfying when a method or plan you set out for a program actually works as intended.
The Spraycan emits particles, which damage spiders over time.
You can also dual-wield them to annihilate the spiders with white, venomous gas. >:>
It's so good. YEAH.
Hmm... not sure what's next. I'll need to check in with the team.